refactor: request handler 클래스 분리
